Float32 for ops

+++ binoculars_llama/metrics.py
@@ -9,12 +9,12 @@ def perplexity(input_ids: np.ndarray, logits: np.ndarray, attention_mask: np.nda
     input_ids = torch.tensor(input_ids, dtype=torch.long)
     attention_mask = torch.tensor(attention_mask, dtype=torch.long)
     ce_loss_fn = torch.nn.CrossEntropyLoss(reduction="none")
-    shifted_logits = logits[..., :-1, :]
+    shifted_logits = logits[..., :-1, :].float()
     shifted_labels = input_ids[..., 1:]
     shifted_attention_mask = attention_mask[..., 1:]
 
     min_len = min(shifted_logits.shape[1], shifted_labels.shape[1], shifted_attention_mask.shape[1])
-    shifted_logits = shifted_logits[:, :min_len, :]
+    shifted_logits = shifted_logits[:, :min_len, :].float()
     shifted_labels = shifted_labels[:, :min_len]
     shifted_attention_mask = shifted_attention_mask[:, :min_len]
 
@@ -25,7 +25,7 @@ def perplexity(input_ids: np.ndarray, logits: np.ndarray, attention_mask: np.nda
         for tid in control_token_ids:
             mask = mask * (shifted_labels != tid)
     ppl = (ce_loss_fn(shifted_logits.transpose(1, 2), shifted_labels) * mask).sum(1) / mask.sum(1)
-    ppl = ppl.to("cpu").float().numpy()
+    ppl = ppl.to("cpu").float().numpy().astype(np.float32)
     return ppl
 
 
@@ -39,8 +39,8 @@ def entropy(p_logits, q_logits, input_ids, pad_token_id, attention_mask=None, co
         mask = torch.tensor(attention_mask, dtype=torch.long)
 
     min_len = min(p_logits.shape[1], q_logits.shape[1], input_ids.shape[1], mask.shape[1])
-    p_logits = p_logits[:, :min_len, :]
-    q_logits = q_logits[:, :min_len, :]
+    p_logits = p_logits[:, :min_len, :].float()
+    q_logits = q_logits[:, :min_len, :].float()
     input_ids = input_ids[:, :min_len]
     mask = mask[:, :min_len]
 
@@ -48,9 +48,9 @@ def entropy(p_logits, q_logits, input_ids, pad_token_id, attention_mask=None, co
         for tid in control_token_ids:
             mask = mask * (input_ids != tid)
     vocab_size = p_logits.shape[-1]
-    p_proba = softmax_fn(p_logits).view(-1, vocab_size)
-    q_scores = q_logits.view(-1, vocab_size)
-    ce = ce_loss_fn(input=q_scores, target=p_proba).view(input_ids.shape[0], -1)
-    padding_mask = mask
-    agg_ce = (((ce * padding_mask).sum(1) / padding_mask.sum(1)).to("cpu").float().numpy())
+    p_proba = softmax_fn(p_logits).view(-1, vocab_size).float()
+    q_scores = q_logits.view(-1, vocab_size).float()
+    ce = ce_loss_fn(input=q_scores, target=p_proba).view(input_ids.shape[0], -1).float()
+    padding_mask = mask.float()
+    agg_ce = (((ce * padding_mask).sum(1) / padding_mask.sum(1)).to("cpu").float().numpy().astype(np.float32))
     return agg_ce
